+++ spec/baby_squeel/nodes/node_spec.rb
@@ -9,7 +9,6 @@ describe BabySqueel::Nodes::Node do
   describe 'included modules' do
     subject { described_class.ancestors }
     specify { is_expected.to include(BabySqueel::Operators::Equality) }
-    specify { is_expected.to include(BabySqueel::Operators::Generic) }
   end
 
   it 'extends any node with math' do