+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- remove DSL#\_
 - remove op, quoted, sql
 - remove BabySqueel.configure and BabySqueel[]
 - remove BabySqueel::ActiveRecord::Calculations (plucking, averaging, counting, maximizing, minimizing, summing)